+++ packages/rubyk-cli/src/cli.ts
@@ -1,5 +1,5 @@
 import { build } from 'gluegun'
-import  rubyk from './commands/Rubyk'
+import rubyk from './commands/Rubyk'
 import { Command } from 'gluegun/build/types/domain/command'
 
 async function run(argv) {
@@ -14,13 +14,12 @@ async function run(argv) {
     .version()
     .create()
 
-    
   cli.defaultCommand = rubyk as Command
   const toolbox = await cli.run(argv)
 
   return toolbox
 }
 
-module.exports = { 
-  run, 
+module.exports = {
+  run,
 }