+++ packages/rubyk-cli/src/extensions/CreateFile.ts
@@ -1,17 +1,22 @@
-import { GluegunToolbox } from "gluegun";
+import { GluegunToolbox } from 'gluegun'
 
 module.exports = (toolbox: GluegunToolbox) => {
   const { template, filesystem } = toolbox
 
-  async function createFile(folder: string, fileName: string, content: string, contentProps?: {[x: string]: any}) {
+  async function createFile(
+    folder: string,
+    fileName: string,
+    content: string,
+    contentProps?: { [x: string]: unknown },
+  ) {
     await template.generate({
       directory: filesystem.path(
-        filesystem.cwd(), 
-        'node_modules/@rubykgen/rubyk-cli/build/templates/src/templates', 
+        filesystem.cwd(),
+        'node_modules/@rubykgen/rubyk-cli/build/templates/src/templates',
       ),
       template: content,
       target: `${folder}/${fileName}`,
-      props: contentProps || {}
+      props: contentProps || {},
     })
   }